requirements.txt
streamlit
pytesseract
pdfplumber
python-docx
langchain
chromadb
openai
pandas
pillow
PyPDF2
pdf2image
google-generativeai
reportlab
fpdf
flask
python-dotenv
google-genai
PyPDF2
python-dotenv